backend/requirements.txt
Django==2.1.1
pytz==2018.9
mysqlclient
django_mysql
djangorestframework
djangorestframework-jwt
